requirements.txt
appdirs==1.4.2
Django==1.10.2
django-hosts==2.0
djangorestframework==3.6.2
gunicorn==19.7.0
packaging==16.8
pyparsing==2.1.10
six==1.10.0
whitenoise==3.3.0
